+++ sdk/storage/azure-storage-files-datalake/src/datalake_path_client.cpp
@@ -345,8 +345,6 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     ret.HttpHeaders.ContentType = std::move(result->HttpHeaders.ContentType);
     ret.IsServerEncrypted = result->IsServerEncrypted;
     ret.EncryptionKeySha256 = std::move(result->EncryptionKeySha256);
-    ret.IsAccessTierInferred = std::move(result->IsAccessTierInferred);
-    ret.AccessTierChangedOn = std::move(result->AccessTierChangedOn);
     ret.CopyId = std::move(result->CopyId);
     ret.CopySource = std::move(result->CopySource);
     ret.CopyStatus = std::move(result->CopyStatus);
@@ -371,7 +369,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
       const Azure::Core::Context& context) const
   {
     _detail::DataLakeRestClient::Path::GetPropertiesOptions protocolLayerOptions;
-    protocolLayerOptions.Action = Models::PathGetPropertiesAction::GetAccessControl;
+    protocolLayerOptions.Action = _detail::PathGetPropertiesAction::GetAccessControl;
     protocolLayerOptions.LeaseIdOptional = options.AccessConditions.LeaseId;
     protocolLayerOptions.IfMatch = options.AccessConditions.IfMatch;
     protocolLayerOptions.IfNoneMatch = options.AccessConditions.IfNoneMatch;
@@ -428,7 +426,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
 
   Azure::Response<Models::SetPathAccessControlListRecursiveSinglePageResult>
   DataLakePathClient::SetAccessControlListRecursiveSinglePageInternal(
-      Models::PathSetAccessControlRecursiveMode mode,
+      _detail::PathSetAccessControlRecursiveMode mode,
       const std::vector<Models::Acl>& acls,
       const SetPathAccessControlListRecursiveSinglePageOptions& options,
       const Azure::Core::Context& context) const